chore(reports): rename legacy view/serializer modules
Also rename reports package to report, following the trend of
all other packages having their name in singular.
