+++ quipucords/api/aggregate_report/model.py
@@ -29,7 +29,7 @@ from math import ceil
 
 from api.deployments_report.model import Product, SystemFingerprint
 from api.inspectresult.model import InspectResult
-from api.reports.model import Report
+from api.report.model import Report
 from constants import DataSources
 from fingerprinter import jboss_eap, jboss_web_server
 from utils.datetime import average_date
