+++ quipucords/api/report/reports_gzip_renderer.py
